extract common util
