+++ src/victory-primitives/curve.js
@@ -1,5 +1,6 @@
 import React, { PropTypes } from "react";
 import Helpers from "../victory-util/helpers";
+import Utils from "./utils";
 import { assign, isEqual } from "lodash";
 import * as d3Shape from "d3-shape";
 
@@ -56,41 +57,11 @@ export default class Curve extends React.Component {
   }
 
   getDataSegments(data) {
-    return this.splitArray(data, (datum) => {
+    return Utils.splitArray(data, (datum) => {
       const yDatum = datum.y1 !== undefined ? datum._y1 : datum._y;
 
       return yDatum === null || typeof yDatum === "undefined";
     }).filter((segment) => segment.length > 1);
-  }
-
-  /**
-   * Split array into subarrays using a delimiter function. Items qualifying as
-   * delimiters are excluded from the subarrays. Functions similarly to String.split
-   *
-   * Example:
-   * const array = [1, 2, 3, "omit", 4, 5, "omit", 6]
-   * splitArray(array, (item) => item === "omit");
-   * => [[1, 2, 3], [4, 5], [6]]
-   *
-   * @param {Array}    array        An array of items
-   * @param {Function} delimiterFn  A function indicating values to be used as delimiters
-   * @returns {Object}              Array of subarrays
-   */
-  splitArray(array, delimiterFn) {
-    let segmentStartIndex = 0;
-    const segments = array.reduce((memo, item, index) => {
-      if (delimiterFn(item)) {
-        memo = memo.concat([array.slice(segmentStartIndex, index)]);
-        segmentStartIndex = index + 1;
-      } else if (index === array.length - 1) {
-        memo = memo.concat([array.slice(segmentStartIndex, array.length)]);
-      }
-      return memo;
-    }, []);
-
-    return segments.filter((segment) => {
-      return Array.isArray(segment) && segment.length > 0;
-    });
   }
 
   toNewName(interpolation) {
